+++ ids/__init__.py
@@ -2,3 +2,4 @@ from ids.blocks import *
 from ids.effects import *
 from ids.entities import *
 from ids.items import *
+from ids.attributes import *