+++ ids/attributes.py
@@ -1,0 +1,37 @@
+class Attribute: "Entity Attribute"
+
+armor: Attribute
+armor_toughness: Attribute
+attack_damage: Attribute
+attack_knockback: Attribute
+attack_speed: Attribute
+block_break_speed: Attribute
+block_interaction_range: Attribute
+burning_time: Attribute
+camera_distance: Attribute
+entity_interaction_range: Attribute
+explosion_knockback_resistance: Attribute
+fall_damage_multiplier: Attribute
+flying_speed: Attribute
+follow_range: Attribute
+gravity: Attribute
+jump_strength: Attribute
+knockback_resistance: Attribute
+luck: Attribute
+max_absorption: Attribute
+max_health: Attribute
+mining_efficiency: Attribute
+movement_efficiency: Attribute
+movement_speed: Attribute
+oxygen_bonus: Attribute
+safe_fall_distance: Attribute
+scale: Attribute
+spawn_reinforcements: Attribute
+sneaking_speed: Attribute
+step_height: Attribute
+submerged_mining_speed: Attribute
+sweeping_damage_ratio: Attribute
+tempt_range: Attribute
+water_movement_efficiency: Attribute
+waypoint_receive_range: Attribute
+waypoint_transmit_range: Attribute